requirements.txt
# Core Backend
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Data Science & ML
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
shap==0.42.1

# Frontend & Visualization
streamlit>=1.28.0
plotly>=5.17.0
folium>=0.14.0
streamlit-folium>=0.13.0

# APIs & External Services
requests>=2.31.0
httpx>=0.24.0
twilio>=8.10.0
sendgrid>=6.10.0

# Development & Utilities
python-dateutil>=2.8.0
joblib>=1.3.0

# Weather API
aiohttp>=3.8.0

# Database
asyncpg>=0.27.0